Users view refactoring - move backend call to the parent component (view) - split UserList component
