+++ viv-wallet-app-vue/src/components/UserList.vue
@@ -1,14 +1,8 @@
 <template>
     <div>
-        <section v-if="errored">
-            <p>We're sorry, we're not able to retrieve this information at the moment, please try back later</p>
-        </section>
-        <section v-else>
-            <div v-if="loading">Loading...</div>
-            <ol v-else>
-                <li v-for="user in users" :key="user.id">{{ user.login }}</li>
-            </ol>
-        </section>
+        <ol>
+            <UserItem v-for="user in users" :key="user.id" v-bind:user="user" />
+        </ol>
     </div>
 </template>
 
@@ -16,25 +10,13 @@
 import { Component, Prop, Vue } from "vue-property-decorator";
 import axios from "axios";
 import { User } from "../models/user";
-import { BACKEND_BASE_URL, REQUEST_TIMEOUT_MS } from "../config/constants";
+import UserItem from "@/components/UserItem.vue";
 
-@Component
+@Component({
+    components: { UserItem }
+})
 export default class UserList extends Vue {
-    users: User[] = [];
-    loading = true;
-    errored = false;
-
-    async mounted() {
-        try {
-            var response = await axios.get<User[]>(`${BACKEND_BASE_URL}/users`, { timeout: REQUEST_TIMEOUT_MS });
-            this.users = response.data;
-        } catch (ex) {
-            console.error(ex);
-            this.errored = true;
-        } finally {
-            this.loading = false;
-        }
-    }
+    @Prop() users!: User[];
 }
 </script>
